Updated Admin
Added the ability for admin to edit, delete and add new users. Admin are able to view images and delete an image. When an image is deleted it updates it saying that it has been deleted and why it has been deleted. The super user is able to edit the score of the user, allowing them to "win" the challenge.